Make MultiLangConfig immutable
